+++ contact/tests/test_forms.py
@@ -308,3 +308,11 @@ def test_selling_online_overseas_business_invalid_form():
     )
     assert form.is_valid() is False
     assert form.errors == {'company_number': ['This field is required.']}
+
+
+@pytest.mark.parametrize('value', (True, False,))
+def test_routing_forms_capital_invest_feature_flag(value, feature_flags):
+    feature_flags['CAPITAL_INVEST_CONTACT_IN_TRIAGE_ON'] = value
+    choices = forms.InternationalRoutingForm().fields['choice'].choices
+
+    assert any(value == constants.CAPITAL_INVEST for value, label in choices) is value